More intelligent scaffolding

+++ lib/mix/operately.ex
@@ -2,9 +2,13 @@ defmodule Mix.Operately do
   # Convinience function for mix tasks
 
   def generate_file(path, generator) do
-    IO.puts "#{IO.ANSI.green()}Generating#{IO.ANSI.reset()} #{path}"
-
-    File.write!(path, generator.(path))
+    if File.exists?(path) do
+      IO.puts "#{IO.ANSI.green()}Aborting#{IO.ANSI.reset()} #{path} already exists"
+      System.halt(1)
+    else
+      IO.puts "#{IO.ANSI.green()}Generating#{IO.ANSI.reset()} #{path}"
+      File.write!(path, generator.(path))
+    end
   end
 
   def indent(lines, spaces) do
